refactor(rc-2): split lib/storage.ts into purpose-specific modules under lib/storage/
